[MARKENG-2177] LeftNav & BreadCrumbs, create removePrefix helper function, use that in nav and breadcrumbs for determining current url
